refactor!: rename `shared_math` into `math`
