+++ twenty-first/Cargo.toml
@@ -33,7 +33,7 @@ features = ["precommit-hook", "run-cargo-clippy", "run-cargo-fmt"]
 
 [dependencies]
 arbitrary = { version = "1", features = ["derive"] }
-bfieldcodec_derive = "0.6"
+bfieldcodec_derive = "0.7"
 blake3 = "1.5.0"
 get-size = { version = "^0.1.4", features = ["derive"] }
 hashbrown = "0.14"
