+++ twenty-first/benches/merkle_tree.rs
@@ -5,9 +5,9 @@ use criterion::criterion_main;
 use criterion::BenchmarkId;
 use criterion::Criterion;
 
-use twenty_first::shared_math::digest::Digest;
-use twenty_first::shared_math::other::random_elements;
-use twenty_first::shared_math::tip5::Tip5;
+use twenty_first::math::digest::Digest;
+use twenty_first::math::other::random_elements;
+use twenty_first::math::tip5::Tip5;
 use twenty_first::util_types::merkle_tree::CpuParallel;
 use twenty_first::util_types::merkle_tree::MerkleTree;
