+++ twenty-first/benches/merkle_tree_authenticate.rs
@@ -2,8 +2,8 @@ use criterion::*;
 use rand::rngs::StdRng;
 use rand::*;
 
-use twenty_first::shared_math::digest::Digest;
-use twenty_first::shared_math::tip5::Tip5;
+use twenty_first::math::digest::Digest;
+use twenty_first::math::tip5::Tip5;
 use twenty_first::util_types::algebraic_hasher::AlgebraicHasher;
 use twenty_first::util_types::merkle_tree::*;
