+++ twenty-first/src/math/b_field_element.rs
@@ -28,11 +28,11 @@ use serde::Serialize;
 use serde::Serializer;
 
 use crate::error::ParseBFieldElementError;
-use crate::shared_math::traits::CyclicGroupGenerator;
-use crate::shared_math::traits::FiniteField;
-use crate::shared_math::traits::ModPowU32;
-use crate::shared_math::traits::ModPowU64;
-use crate::shared_math::traits::New;
+use crate::math::traits::CyclicGroupGenerator;
+use crate::math::traits::FiniteField;
+use crate::math::traits::ModPowU32;
+use crate::math::traits::ModPowU64;
+use crate::math::traits::New;
 
 use super::traits::Inverse;
 use super::traits::PrimitiveRootOfUnity;
@@ -660,9 +660,9 @@ mod b_prime_field_element_test {
     use rand::thread_rng;
     use test_strategy::proptest;
 
-    use crate::shared_math::b_field_element::*;
-    use crate::shared_math::other::random_elements;
-    use crate::shared_math::polynomial::Polynomial;
+    use crate::math::b_field_element::*;
+    use crate::math::other::random_elements;
+    use crate::math::polynomial::Polynomial;
 
     impl proptest::arbitrary::Arbitrary for BFieldElement {
         type Parameters = ();
