+++ twenty-first/src/prelude.rs
@@ -2,17 +2,17 @@ pub use crate::amount::u32s::U32s;
 pub use crate::bfe;
 pub use crate::bfe_array;
 pub use crate::bfe_vec;
-pub use crate::shared_math::b_field_element;
-pub use crate::shared_math::b_field_element::BFieldElement;
-pub use crate::shared_math::bfield_codec::BFieldCodec;
-pub use crate::shared_math::polynomial::Polynomial;
-pub use crate::shared_math::tip5;
-pub use crate::shared_math::tip5::Digest;
-pub use crate::shared_math::tip5::Tip5;
-pub use crate::shared_math::traits::Inverse;
-pub use crate::shared_math::traits::ModPowU32;
-pub use crate::shared_math::x_field_element;
-pub use crate::shared_math::x_field_element::XFieldElement;
+pub use crate::math::b_field_element;
+pub use crate::math::b_field_element::BFieldElement;
+pub use crate::math::bfield_codec::BFieldCodec;
+pub use crate::math::polynomial::Polynomial;
+pub use crate::math::tip5;
+pub use crate::math::tip5::Digest;
+pub use crate::math::tip5::Tip5;
+pub use crate::math::traits::Inverse;
+pub use crate::math::traits::ModPowU32;
+pub use crate::math::x_field_element;
+pub use crate::math::x_field_element::XFieldElement;
 pub use crate::util_types::algebraic_hasher::AlgebraicHasher;
 pub use crate::util_types::algebraic_hasher::Sponge;
 pub use crate::util_types::merkle_tree::CpuParallel;
