+++ twenty-first/src/util_types/merkle_tree_maker.rs
@@ -1,4 +1,4 @@
-use crate::shared_math::digest::Digest;
+use crate::math::digest::Digest;
 use crate::util_types::algebraic_hasher::AlgebraicHasher;
 use crate::util_types::merkle_tree::*;
