+++ twenty-first/src/util_types/mmr/shared_basic.rs
@@ -1,5 +1,5 @@
 use super::mmr_membership_proof::MmrMembershipProof;
-use crate::shared_math::digest::Digest;
+use crate::math::digest::Digest;
 use crate::util_types::algebraic_hasher::AlgebraicHasher;
 
 #[inline]
