+++ twenty-first/trybuild/missing_field_attribute.rs
@@ -1,5 +1,5 @@
 use twenty_first;
-use twenty_first::shared_math::bfield_codec::BFieldCodec;
+use twenty_first::math::bfield_codec::BFieldCodec;
 
 #[derive(BFieldCodec)]
 struct MyStruct {
